% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDJac.R
\name{CDJac}
\alias{CDJac}
\title{CDJac}
\usage{
CDJac(f, x0, ...)
}
\arguments{
\item{f}{pointer to the routine that returns the m-vector f(x)}

\item{x0}{m vector x0, the point at which the derivatives are to be evaluated}

\item{...}{any additional arguments passed to \code{f}.}
}
\value{
Jac n by m matrix of parital derivatives
}
\description{
computes a central difference approximation of the Jacobian
matrix of a system of n non-linear functions y_i=f^i(x), where
x is a column vector of dimension m.
}
\details{
algorithm: based on (A.2.8) in Heer and Maussner, see also Dennis and Schnabel (1983),
Algorithm A5.6.4.
}
