% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckBounds2.R
\name{CheckBounds2}
\alias{CheckBounds2}
\title{CheckBounds2}
\usage{
CheckBounds2(x, dx, bounds)
}
\arguments{
\item{x}{k times 1 vector, the given value of x0}

\item{dx}{k times 1 vector, the Newton step from x0 to x1}

\item{bounds}{k times 2 matrix, the first (second) colum store the lower (upper) bounds of x}
}
\value{
s a scalar
}
\description{
Finds smallest s so that x+s*dx is within bounds
}
