% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixVMN2.R
\name{FixVMN2}
\alias{FixVMN2}
\title{FixvMN2}
\usage{
FixVMN2(
  x0,
  bounds,
  f,
  stopc = 1e-07,
  pTol = .Machine$double.eps^(2/3),
  MNR_CDJac = TRUE,
  MNR_Global = FALSE,
  ...
)
}
\arguments{
\item{x0}{= k times 1 vector of starting values}

\item{bounds}{k timex 2 vector of upper and lower bounds on x}

\item{f}{Pointer the vector valued function F(x), whose
zero, F(x1)=0, is to be computed.}

\item{stopc}{stopping criterium}

\item{pTol}{parameter convergence criterium}

\item{MNR_CDJac}{wheather using \code{CDJac} computing Jacobian matrix, with fault value is TRUE.}

\item{MNR_Global}{wheather using line search, with fault value is FALSE.}

\item{...}{any additional arguments passed to \code{f}.

@return a list including x1 and crit. x1 is k times 1 vector, the approximate solution to F(x1)=0.
\code{crit} is 1 vector, where
\code{crit[1]=0} : normal termination
=1 : function evaluation failed,
=2 : no further decrease in function value possible,
=3 : maximum number of iterations exceedes
\code{crit[2]}   : termination criterion: max(abs(f(x)))
\code{crit[3]}   : the maximum relative change of x between the last two iterations
\code{crit[4]}   : f(x)'f(x)/2
\code{crit[5]}   : number of iterations}
}
\description{
Solves a system of non-linear equations using a modified Newton Method
}
