% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinStep.R
\name{MinStep}
\alias{MinStep}
\title{Min Step}
\usage{
MinStep(x, dx, typx, pTol)
}
\arguments{
\item{x}{n times 1 vector}

\item{dx}{n times 1 vector, change in x}

\item{typx}{n times 1 vector, typical elements of x}

\item{pTol}{scalar}
}
\value{
mstep scalar
}
\description{
Compute the minimal step size so that x1=x0+mstep*dx=0
(in terms of the parameter tolerance criterium pTol)
}
