% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNRStep.R
\name{MNRStep}
\alias{MNRStep}
\title{MNRStep}
\usage{
MNRStep(x0, dx0, dg, pTol, f, ...)
}
\arguments{
\item{x0}{n times 1 vector, the initial point}

\item{dx0}{n times 1 vector, the Newton direction}

\item{dg}{1 times n vector, the gradient of (1/2)(f'f) at x_0}

\item{pTol}{scalar, parameter convergence criterium}

\item{f}{pointer to the function whose zero x solves the system of equations}

\item{...}{any additional arguments passed to \code{f}.}
}
\value{
s  admissible stepsize
}
\description{
Find the step size s so that the Newton-Raphson algorithm
always moves in the direction of a (local) minimum of (1/2)(f(x)'f(x))
}
