% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParTest.R
\name{ParTest}
\alias{ParTest}
\title{Parameters Test}
\usage{
ParTest(x, dx, typx)
}
\arguments{
\item{x}{:= n times 1 vector}

\item{dx}{:= n times 1 vector, change in x}

\item{typx}{:= n times 1 vector, typical elements of x}
}
\value{
crit := scalar
}
\description{
Compute relative change in x
}
